# docs/swagger.yaml
basePath: /
definitions:
...
  hydration.Entry:
    properties:
      created_at:
        type: string
      id:
        type: string
      logged_at:
        type: string
      note:
        type: string
      quantity_ml:
        type: number
      updated_at:
        type: string
      workout_id:
        type: string
    type: object
  hydration.createRequest:
    properties:
      logged_at:
        type: string
      note:
        type: string
      quantity_ml:
        type: number
      workout_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Personal nutrition tracking REST API. All endpoints under /products,
    /meals, and /summary require a bearer token.
  title: Nutrition API
  version: 0.1.0
paths:
...
  /garmin/activity/{activity_id}:
    delete:
      description: Forwards to the bridge, which deletes the Garmin activity. An already-absent
        activity is a no-op success.
      parameters:
      - description: Garmin activity id
        in: path
        name: activity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the bridge response verbatim
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable | garmin_error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a Garmin activity (idempotent)
      tags:
      - garmin
    patch:
      consumes:
      - application/json
      parameters:
      - description: Garmin activity id
        in: path
        name: activity_id
        required: true
        type: string
      - description: '{ name }'
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: the bridge response verbatim
          schema:
            additionalProperties: true
            type: object
        "400":
          description: name_required
          schema:
            additionalProperties:
              type: string
            type: object
        "502":
          description: garmin_bridge_unreachable | garmin_error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Rename a Garmin activity
      tags:
      - garmin
  /garmin/activity/{activity_id}/export:
    get:
      description: Forwards to the bridge, returning `{activity_id, format, filename,
        content_base64}` verbatim — the file bytes base64-encoded in JSON. `format`
        defaults to `fit`. Upload is out of scope.
      parameters:
      - description: Garmin activity id
        in: path
        name: activity_id
        required: true
        type: string
      - description: fit (default) | gpx | tcx | kml | csv
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ activity_id, format, filename, content_base64 }'
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable | garmin_error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export an activity's FIT/GPX/TCX file (base64 envelope)
      tags:
      - garmin
  /garmin/activity/{activity_id}/gear:
    get:
      parameters:
      - description: Garmin activity id
        in: path
        name: activity_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the bridge activity-gear response verbatim
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Read the gear linked to a Garmin activity
      tags:
      - garmin
  /garmin/activity/upload:
    post:
      consumes:
      - application/json
      description: Forwards a base64-wrapped FIT payload `{filename, content_base64}`
        to the bridge, which uploads it via Garmin. Writes FROM this API TO Garmin
        — explicit-request only.
      parameters:
      - description: '{ filename, content_base64 }'
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: the bridge upload response verbatim
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable | garmin_error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload a FIT activity to Garmin (opt-in write)
      tags:
      - garmin
  /garmin/calendar:
    get:
      parameters:
      - description: Inclusive lower bound YYYY-MM-DD
        in: query
        name: from
        required: true
        type: string
      - description: Inclusive upper bound YYYY-MM-DD
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: the bridge calendar response verbatim
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Read the Garmin calendar through the bridge
      tags:
      - garmin
  /garmin/hydration:
    post:
      consumes:
      - application/json
      description: Forwards `{value_ml, date}` to the bridge, which sets/replaces
        that day's hydration on Garmin. The only write FROM us TO Garmin — invoked
        deliberately, nothing pushes automatically.
      parameters:
      - description: '{ value_ml, date }'
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: the bridge response verbatim
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable | garmin_error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Push logged hydration back to Garmin (opt-in write)
      tags:
      - garmin
  /garmin/login:
    post:
      description: 'Forwards to the garmin-bridge''s POST /login, returning its response
        verbatim (e.g. `{"needs_mfa": true}`). Carries NO credentials — the bridge
        reads them from its own configuration. Returns 503 garmin_disabled when GARMIN_BRIDGE_URL
        is unset. Any authenticated identity may call it.'
      produces:
      - application/json
      responses:
        "200":
          description: Bridge response verbatim (e.g. {needs_mfa:true} or {logged_in:true})
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start the Garmin bridge login (proxy)
      tags:
      - garmin
  /garmin/login/mfa:
    post:
      consumes:
      - application/json
      description: Forwards the supplied 6-digit code to the garmin-bridge's POST
        /login/mfa, returning its success/error response verbatim. On success the
        bridge persists the minted token to the backend; it is never returned here.
        Returns 503 garmin_disabled when GARMIN_BRIDGE_URL is unset.
      parameters:
      - description: '{ \'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Bridge response verbatim (e.g. {logged_in:true})
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Submit the Garmin bridge MFA code (proxy)
      tags:
      - garmin
...
  /garmin/workout/{garmin_workout_id}/download:
    get:
      description: 'The structured-workout analogue of the activity export: forwards
        to the bridge and returns `{garmin_workout_id, format, filename, content_base64}`
        verbatim. `format` defaults to `fit`.'
      parameters:
      - description: Garmin workout object id
        in: path
        name: garmin_workout_id
        required: true
        type: string
      - description: fit (default) | …
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ garmin_workout_id, format, filename, content_base64 }'
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable | garmin_error
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download a structured workout's FIT blob (base64 envelope)
      tags:
      - garmin
...
  /garmin/workouts:
    get:
      parameters:
      - description: Pagination start offset (passthrough)
        in: query
        name: start
        type: integer
      - description: Pagination limit (passthrough)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: the bridge workout-library response verbatim
          schema:
            additionalProperties: true
            type: object
        "502":
          description: garmin_bridge_unreachable
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: garmin_disabled
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List the Garmin workout library through the bridge
      tags:
      - garmin
...
  /hydration:
    get:
      parameters:
      - description: Inclusive RFC3339 lower bound
        in: query
        name: from
        required: true
        type: string
      - description: Exclusive RFC3339 upper bound
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{ entries: [...] }'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: window_required | window_invalid | range_too_large
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List hydration entries in a window
      tags:
      - hydration
    post:
      consumes:
      - application/json
      description: Records a volume of fluid drunk at a moment in time. The optional
        `note` carries beverage context (e.g. "water", "coffee", "electrolytes").
      parameters:
      - description: Optional client-supplied idempotency key
        in: header
        name: Idempotency-Key
        type: string
      - description: Hydration entry
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/hydration.createRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hydration.Entry'
        "400":
          description: quantity_ml_invalid | logged_at_invalid | logged_at_too_far_future
            | note_too_long
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Log a hydration entry
      tags:
      - hydration
...
securityDefinitions:
  BearerAuth:
    description: Static bearer token. Use the format `Bearer <token>` with either
      MOBILE_API_TOKEN or AGENT_API_TOKEN.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"